# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "URL Shortener",
        "title": "Go Shorten API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "leachiM2k",
            "url": "https://github.com/leachim2k/go-shorten",
            "email": "leachiM2k@leachiM2k.de"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/shorten": {
            "get": {
                "description": "Get all shorts for an user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all user shorts",
                "operationId": "readAll",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dataservice.Entity"
                            }
                        }
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shorten/": {
            "post": {
                "description": "Create a new short. Create random code if not specified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Add a new short",
                "operationId": "create",
                "parameters": [
                    {
                        "description": "Create Request",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dataservice.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataservice.Entity"
                        }
                    }
                }
            }
        },
        "/shorten/handle/{code}": {
            "get": {
                "description": "Return the right link for short code or \"not found\" if expired, not started or max count was reached",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "summary": "Handle a short",
                "operationId": "handle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Link to follow",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shorten/{code}": {
            "get": {
                "description": "Get all stored information for a specified short",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get short's info",
                "operationId": "read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataservice.Entity"
                        }
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates several fields of a short, while maintaining count, owner and creation date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a short",
                "operationId": "update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Request",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dataservice.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataservice.Entity"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a short",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a short",
                "operationId": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/shorten/{code}/stats": {
            "get": {
                "description": "Get all stats for a code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all stats for a code",
                "operationId": "readStats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dataservice.StatEntity"
                            }
                        }
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/url/meta": {
            "get": {
                "description": "Get all stats for a code",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get all stats for a code",
                "operationId": "getUrlMetaData",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to get meta information from",
                        "name": "url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dataservice.HTMLMeta"
                        }
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/url/qrcode": {
            "get": {
                "description": "Generate QR Code for an URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Generate QR Code",
                "operationId": "getQrCodeForUrl",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL for QR Code",
                        "name": "url",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "image format (svg or png)",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "for PNG: width of image",
                        "name": "width",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "for PNG: height of image",
                        "name": "height",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "fail",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dataservice.CreateRequest": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "maxCount": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "dataservice.Entity": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "code": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "maxCount": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "dataservice.HTMLMeta": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "site_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dataservice.StatEntity": {
            "type": "object",
            "properties": {
                "clientIP": {
                    "type": "string"
                },
                "referer": {
                    "type": "string"
                },
                "shortenerID": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                }
            }
        },
        "dataservice.UpdateRequest": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "additionalProperties": true
                },
                "description": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "maxCount": {
                    "type": "integer"
                },
                "owner": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        }
    },
    "x-extension-openapi": {
        "example": "value on a json format"
    }
}